+++ CHANGELOG.md
@@ -16,3 +16,20 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Update invoice count filtering
 - Renamed invoice primary key from `invoice_id` to `id`
 - Updated invoice check constraints
+
+## [0.7.0] - 2026-##-##
+
+### Added
+- Service layer for customers and invoices
+- Converted from db sqlite.row to dictionaries in services  
+
+### Changed
+- Modify cli tests to service changes
+
+### Considerations
+- Convert data from dict to dataclass or pydantic model
+- Look at and maybe seperate tests where id is 9999
+because -9999 is a validationerror and 9999 is notfounderror
+
+### Need to Change
+- create database services for create, drop, delete